PDFBOX-2576: fix cast from int multiplication to long
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1651951 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/io/RandomAccessBuffer.java
@@ -265,7 +265,7 @@ public class RandomAccessBuffer implements RandomAccess, Closeable, Cloneable
                 newOffset += BUFFER_SIZE;
             }
             // are there still some bytes to be written?
-            remainingBytes2Write -= numberOfNewArrays * BUFFER_SIZE;
+            remainingBytes2Write -= numberOfNewArrays * (long) BUFFER_SIZE;
             if (remainingBytes2Write >= 0)
             {
                 expandBuffer();
